Closes #16
