+++ src/components/Navbar.tsx
@@ -2,11 +2,13 @@ import { RootState } from '../state/state';
 import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
 import './styles/Navbar.css';
+import ThemeChanger from './ThemeChanger';
 
 function Navbar() {
   const theme = useSelector((state: RootState) => state.theme.theme);
   return (
     <div className={`navbar ${theme}`}>
+      <ThemeChanger />
       <Link to={'/'}>Homepage</Link>
       <Link to={'/booklisting'}>Books</Link>
       <Link to={'/events'}>Events</Link>
