+++ src/components/styles/Navbar.css
@@ -6,13 +6,13 @@
 }
 
 .navbar.LightTheme {
-    background-color: rgba(134, 134, 134, 0.5);
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+    background-color: rgba(242, 242, 242, 0.6);
+    box-shadow: 0 0 6px rgba(0, 0, 0, 0.3);
 }
 
 .navbar.DarkTheme {
-    box-shadow: 0 0 10px rgba(255, 255, 255, 0.5);
-    background-color: rrgba(222, 222, 222, 0.5);
+    box-shadow: 0 0 3px rgba(255, 255, 255, 0.3);
+    background-color: rgb(003, 003, 003);
 }
 
 .navbar a {
@@ -28,4 +28,3 @@
 .navbar a:hover {
     color: #007bff;
 }
-
