+++ src/pages/AD1.tsx
@@ -1,0 +1,12 @@
+function AD1() {
+  return (
+    <div
+      className="ad1"
+      id="ad1"
+    >
+      Ad 1
+    </div>
+  );
+}
+
+export default AD1;
