+++ src/pages/AD2.tsx
@@ -1,0 +1,12 @@
+function AD2() {
+  return (
+    <div
+      className="ad2"
+      id="ad2"
+    >
+      Ad 2
+    </div>
+  );
+}
+
+export default AD2;
