+++ src/pages/about/About.tsx
@@ -1,3 +1,5 @@
+import './styles/About.css'
+
 function About() {
   return <div></div>;
 }
