+++ src/pages/bookpreview/BookPreview.tsx
@@ -1,3 +1,5 @@
+import './styles/BookPreview.css';
+
 function BookPreview() {
   return <div></div>;
 }
