+++ src/pages/bookreading/BookReading.tsx
@@ -1,3 +1,5 @@
+import './styles/BookReading.css';
+
 function BookReading() {
   return <div></div>;
 }
