+++ src/pages/contact/components/ContactText.tsx
@@ -1,0 +1,42 @@
+function ContactText() {
+  return (
+    <div
+      className="text-section-contact"
+      id="text-section-contact"
+    >
+      <div>
+        <h1>Contact Page</h1>
+        <p>
+          This page is designed to facilitate communication and improve your
+          experience. It consists of three main sections:
+        </p>
+        <ul>
+          <li>
+            <strong>Feedback</strong>: This section is dedicated to your
+            thoughts and ideas. If you have suggestions on how we can improve
+            your experience, or if you have innovative ideas that you believe
+            could enhance our page, please share them here. Your feedback is
+            invaluable in our continuous effort to improve.
+          </li>
+          <li>
+            <strong>Bug Reporting</strong>: If you encounter any technical
+            issues or bugs while using our page, please report them in this
+            section. Your reports are crucial for us to maintain the smooth
+            operation of our site. Please note that this section is solely for
+            reporting bugs and technical issues.
+          </li>
+          <li>
+            <strong>Contact</strong>: This section is for personal contact. If
+            you need to reach out to us for serious reasons, please use this
+            section. We value all our users and are here to assist you. However,
+            we kindly ask that this section be used only for serious inquiries
+            to ensure that urgent matters receive timely attention.
+          </li>
+        </ul>
+        <p>Thank you for your cooperation and understanding.</p>
+      </div>
+    </div>
+  );
+}
+
+export default ContactText;
