+++ src/pages/contact/components/Slider.tsx
