+++ src/pages/events/Events.tsx
@@ -1,3 +1,5 @@
+import './styles/Events.css'
+
 function Events() {
   return <div></div>;
 }
