+++ src/pages/homepage/components/HeroSection.tsx
@@ -1,0 +1,12 @@
+function HeroSection() {
+  return (
+    <div
+      className="hero-section"
+      id="hero-section"
+    >
+      <p className={`hero-text`}>Librarium Imperialis</p>
+    </div>
+  );
+}
+
+export default HeroSection;
